requirements.txt
asgiref==3.5.0
Django==3.2
pytz==2022.1
sqlparse==0.4.2
django-taggit==2.1.0
django-ckeditor==6.4.0